Use size_t instead of int in wait()

+++ rviz_visual_testing_framework/include/rviz_visual_testing_framework/internal/executor.hpp
@@ -39,7 +39,7 @@ class Executor : public QObject
 public:
   Executor();
   void queueAction(std::function<void(void)> action);
-  void wait(int milliseconds_to_wait);
+  void wait(size_t milliseconds_to_wait);
 
 private:
   void increaseTotalDelay();
