+++ rviz_visual_testing_framework/src/internal/executor.cpp
@@ -43,7 +43,7 @@ void Executor::queueAction(std::function<void(void)> action)
   increaseTotalDelay();
 }
 
-void Executor::wait(int milliseconds_to_wait)
+void Executor::wait(size_t milliseconds_to_wait)
 {
   total_delay_ += milliseconds_to_wait;
 }
